move sppo_filters to utils @br_rj_riodejaneiro_onibus_gps
