+++ pipelines/rj_smtr/br_rj_riodejaneiro_onibus_gps/tasks.py
@@ -15,7 +15,8 @@ from pipelines.utils.utils import log
 # SMTR Imports #
 
 from pipelines.rj_smtr.constants import constants
-from pipelines.rj_smtr.utils import log_critical, sppo_filters
+from pipelines.rj_smtr.utils import log_critical
+from pipelines.rj_smtr.br_rj_riodejaneiro_onibus_gps.utils import sppo_filters
 
 # Tasks #
